requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
scikit-learn
sentence-transformers
passlib[bcrypt]
python-jose
joblib
